+++ dask_match/__init__.py
@@ -1,2 +1,2 @@
 from dask_match.expr import *
-from dask_match.collection import *
+from dask_match.dataframe import *
