--- dask_match/collection/__init__.py
@@ -1,1 +1,0 @@
-from dask_match.collection.core import *
