+++ dask_match/expr/__init__.py
@@ -1,2 +1,1 @@
-from dask_match.expr.core import *
-from dask_match.expr.io import *
+from dask_match.expr.dataframe import *
