--- dask_match/expr/io/__init__.py
@@ -1,3 +1,0 @@
-from dask_match.expr.io.csv import *
-from dask_match.expr.io.io import *
-from dask_match.expr.io.parquet import *
